requirements.txt
fastapi
uvicorn
requests
openai
pytesseract
pillow
beautifulsoup4

